Rename mypy.experiments -> mypy.state
