+++ mypy/semanal_pass1.py
@@ -19,7 +19,7 @@ those with platform checks.
 
 from typing import List, Tuple
 
-from mypy import experiments
+from mypy import state
 from mypy.nodes import (
     MypyFile, SymbolTable, SymbolTableNode, Var, Block, AssignmentStmt, FuncDef, Decorator,
     ClassDef, TypeInfo, ImportFrom, Import, ImportAll, IfStmt, WhileStmt, ForStmt, WithStmt,
@@ -76,7 +76,7 @@ class SemanticAnalyzerPass1(NodeVisitor[None]):
 
         defs = file.defs
 
-        with experiments.strict_optional_set(options.strict_optional):
+        with state.strict_optional_set(options.strict_optional):
             # Add implicit definitions of module '__name__' etc.
             for name, t in implicit_module_attrs.items():
                 # unicode docstrings should be accepted in Python 2
